refactor(util): module structure
