+++ cache/Cargo.toml
@@ -24,4 +24,3 @@ twilight-model = "0.10"
 twilight-util = { version = "0.10", features = ["permission-calculator"] }
 url = { version = "2", features = ["serde"] }
 zstd = "0.11"
-
