+++ cache/src/model/channel.rs
@@ -1,4 +1,4 @@
-use raidprotect_util::serde::IdAsU64;
+use raidprotect_model::serde::IdAsU64;
 use serde::{Deserialize, Serialize};
 use serde_with::serde_as;
 use twilight_model::{
