+++ model/src/serde.rs
